+++ packages/plugins/Gfi/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix: Adjust documentation to properly describe optionality of configuration parameters.
 - Fix: Add missing configuration parameters `featureList` and `maxFeatures` to the general documentation and `filterBy` and `format` to `gfi.gfiLayerConfiguration`
 - Fix: Add missing entry of `gfiContentComponent` to `GfiGetters`.
+- Fix: Fix issue rendering properties of a feature if a value is not a string.
 - Refactor: Replace redundant prop-forwarding with `getters`.
 - Refactor: Use core getter `clientWidth` instead of local computed value.
 - Chore: expand on the description to `gfiContentComponent` in the Readme.md.